requirements.txt
asgiref==3.3.4
boto3==1.17.75
botocore==1.20.75
dj-database-url==0.5.0
Django==3.2
django-allauth==0.44.0
django-countries==7.2.1
django-crispy-forms==1.11.2
django-storages==1.11.1
gunicorn==20.1.0
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
sqlparse==0.4.1
stripe==2.56.0
